requirements.txt
faiss-cpu>=1.10.0
sentence-transformers>=2.2.2
numpy>=1.23
torch>=1.12.0
streamlit>=1.30.0